+++ cuda_mode/matmul/src/run_fp16.cuh
@@ -15,4 +15,4 @@ void check_cuda(cudaError_t err, const char* const func, const char* const file,
 
 void cuda_device_info();
 
-void run_kernel_fp16(int kernel_num, int m, int n, int k, half alpha, half *A, half *B, half beta, half *C, half *D, cublasHandle_t handle);
+void run_kernel_fp16(int kernel_num, int m, int n, int k, float alpha, half *A, half *B, float beta, half *C, half *D, cublasHandle_t handle);
